Refactor : 2D 컷씬 재생 함수 ChapterManager로 이동
