+++ Assets/00Scripts/Timeline/Intro/IntroCutScene.cs
@@ -26,12 +26,9 @@ public class IntroCutScene : CutSceneBase
 	[SerializeField] private List<string> inputText;
 	[SerializeField] private float inputDelay = 0.05f;
 	private int curIndex;
-
+	
+	// 2D 애니 컷 씬
 	private Queue<SkeletonGraphic> cutSceneQueue;
-	private bool isInput = false;
-	private bool isInputCheck = false;
-
-	private IEnumerator skeletonCutScene;
 	
 	protected override void Init()
 	{
@@ -47,13 +44,10 @@ public class IntroCutScene : CutSceneBase
 	protected override void EnableCutScene()
 	{
 		FadeManager.Instance.FadeOut(fadeInTime);
-		InputActionManager.Instance.RegisterCallback(InputActionManager.Instance.InputActions.UIBehaviour.ClickUI, InputCheck, true);
 	}
 
 	protected override void DisableCutScene()
 	{
-		InputActionManager.Instance.RemoveCallback(InputActionManager.Instance.InputActions.UIBehaviour.ClickUI, InputCheck);
-		
 		FadeManager.Instance.FadeIn(fadeOutTime, () =>
 		{
 			SceneLoader.Instance.LoadScene(nextSceneName);
@@ -66,70 +60,12 @@ public class IntroCutScene : CutSceneBase
 
 		StartCoroutine(PrintText());
 	}
-
-#if UNITY_EDITOR
-	private void Update()
-	{
-		if (Input.GetKeyDown(KeyCode.F) && isInputCheck == true)
-		{
-			isInput = true;
-		}
-	}
-	#endif
-
+	
 	public void StartSkeletonCutScene()
 	{
-		PauseTimeline();
-		
-		skeletonCutScene = SkeletonCutScene();
-		StartCoroutine(skeletonCutScene);
+		chapterManager.StartSkeletonCutScene(introCutScene, cutSceneQueue);
 	}
-
-	private IEnumerator SkeletonCutScene()
-	{
-		SkeletonGraphic skeleton = cutSceneQueue.Dequeue();
-		int curAniIndex = 0;
-		int maxAniCount = skeleton.Skeleton.Data.Animations.Count;
-
-		while (true)
-		{
-			skeleton.gameObject.SetActive(true);
-			isInputCheck = true;
-			
-			Animation ani = skeleton.Skeleton.Data.Animations.Items[curAniIndex];
-			skeleton.AnimationState.SetAnimation(0, ani, false);
-			
-			while (isInput == false)
-			{
-				yield return null;
-			}
-
-			if (curAniIndex + 1 < maxAniCount)
-			{
-				curAniIndex++;
-			}
-			else
-			{
-				if (cutSceneQueue.Count <= 0)
-				{
-					skeleton.gameObject.SetActive(false);
-					break;
-				}
-				
-				skeleton.gameObject.SetActive(false);
-				skeleton = cutSceneQueue.Dequeue();
-
-				curAniIndex = 0;
-				maxAniCount = skeleton.Skeleton.Data.Animations.Count;
-			}
-			
-			isInput = false;
-		}
-		
-		isInputCheck = isInput = false;
-		introCutScene.Resume();
-	}
-
+	
 	private IEnumerator PrintText()
 	{
 		for (int i = 0; i < inputText[curIndex].Length; ++i)
@@ -142,14 +78,5 @@ public class IntroCutScene : CutSceneBase
 		curIndex++;
 		introCutScene.Resume();
 	}
-
-	private void InputCheck(InputAction.CallbackContext context)
-	{
-		isInput = true;
-	}
 	
-	private void PauseTimeline()
-	{
-		introCutScene.Pause();
-	}
 }
